+++ Dtos/PropertyTraceDto.cs
@@ -5,7 +5,7 @@ namespace RealEstate.API.Dtos
         public string IdPropertyTrace { get; set; } = string.Empty;
         public string DateSale { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
-        public decimal Value { get; set; }
-        public decimal Tax { get; set; }
+        public long Value { get; set; }
+        public long Tax { get; set; }
     }
 }
